test: restructuring tests
