+++ jest.config.js
@@ -3,7 +3,7 @@ module.exports = {
   testEnvironment: 'jsdom',
 
   testMatch: [
-    '**/__tests__/**/*.[jt]s?(x)',
+    '**/tests/**/*.[jt]s?(x)',
     '**/?(*.)+(spec|test).[jt]s?(x)'
   ],
